+++ java/demo/pom.xml
@@ -11,8 +11,6 @@
     <plugins>
       <plugin>
         <artifactId>appengine-maven-plugin</artifactId>
-        <configuration>
-        </configuration>
         <groupId>com.google.cloud.tools</groupId>
         <version>2.4.4</version>
       </plugin>
@@ -37,7 +35,7 @@
     <dependency>
       <artifactId>appengine-api-1.0-sdk</artifactId>
       <groupId>com.google.appengine</groupId>
-      <version>1.9.96</version>
+      <version>2.0.8</version>
     </dependency>
     <dependency>
       <artifactId>servlet-api</artifactId>
