+++ java/demo/src/main/resources/com/google/i18n/phonenumbers.demo/input_form.soy
@@ -1,5 +1,5 @@
-{namespace examples.simple}
-{template helloWorld}
+{namespace demo.input}
+{template inputForm}
 {@param welcomeTitle: string}
 <html>
   <head>
