+++ java/demo/src/main/resources/com/google/i18n/phonenumbers.demo/result.soy
@@ -1,5 +1,5 @@
-{namespace examples.simple}
-{template result}
+{namespace demo.output}
+{template singleNumber}
 {@param phoneNumber: string}
 {@param defaultCountry: string}
 {@param geocodingLocale: string}
