+++ java/demo/src/main/webapp/WEB-INF/appengine-web.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appengine-web-app xmlns="http://appengine.google.com/ns/1.0">
   <application>libphonenumber-hrd</application>
-  <version>1</version>
-  <runtime>java8</runtime>
+  <runtime>java11</runtime>
+  <system-properties>
+    <property name="java.util.logging.config.file"
+        value="WEB-INF/logging.properties"/>
+  </system-properties>
   <threadsafe>true</threadsafe>
 
   <!-- Configure java.util.logging -->
-  <system-properties>
-    <property name="java.util.logging.config.file" value="WEB-INF/logging.properties"/>
-  </system-properties>
+  <version>1</version>
 </appengine-web-app>
